# bpo-46993: Speed up bytearray creation from list and tuple