Speeds up bytearray creation from list and tuples.

Benchmark:
```py
import pyperf

runner = pyperf.Runner()
runner.timeit(name="bench bytearray",
              stmt="bytearray([42]*10000)",)
```

Result:
```console
Mean +- std dev: [base] 74.8 us +- 5.5 us -> [patch] 53.2 us +- 3.3 us: 1.41x faster
```

<!-- issue-number: [bpo-46993](https://bugs.python.org/issue46993) -->
https://bugs.python.org/issue46993
<!-- /issue-number -->